requirements.txt
python-dateutil==2.8.2
six==1.16.0
fs~=2.4.12
pip~=22.0.2
attrs~=21.2.0
wheel~=0.37.1
Brotli~=1.0.9
cryptography~=41.0.4
lxml~=4.8.0
Pillow~=9.0.1
keyring~=23.5.0
setuptools~=59.6.0
future~=0.18.2
prettytable~=3.9.0
pandas~=2.1.1
et-xmlfile~=21.1.0
openpyxl~=23.1.2
pdfquery~=0.4.3
requests~=2.25.1